move MMFeffDb to feff.database.py
